feat: Changed the addressing of the LEDs over to use a matrix
This has been done to allow for the addresses of each of the LEDs to be
mapped to their position in a loop more easily. The downside to this
approach is that all of the addresses for the LEDs will need to be typed
out by hand.
Also shuffled some of the defines and variables around to better group
them in the code.