refactor: Changed up the FPS Accel/Decel
I did this to add a bit of randomness into the FPS rate that was used
each cycle, to try and keep things a bit more interesting for the
viewers.
This would have worked much better with a higher base FPS, as I could
not decrease the fps too much with the current setup, before it
started to look way more choppier than usual.