refactor: Tweaked the random palette picking code, and added more palettes
The prior setup for selecting colours relyed on the programmer to code
in the combinations for each of the colour palettes in the program.
This new method picks two random numbers that are not the same, and
are also both valid indices for a list of all the pre-programmed
palettes.
It then loads the palettes from those indices into the appropriate
place in the hue array for each zoetrope loop, then stops until
the next time it is needed.